langraph_project/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# LangGraph + LangChain
langgraph==0.1.19
langchain==0.2.5
langchain-core==0.2.9
langchain-openai==0.1.13
langchain-community==0.2.5

# RAG / Vector store
faiss-cpu==1.8.0
tiktoken==0.7.0

# Resilience
tenacity==8.3.0

# Observability
langsmith==0.1.77
arize-phoenix>=4.0.0
openinference-instrumentation-langchain>=0.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP client
httpx==0.27.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
